chore: rename serialize to serialization
